Fix concurrency warnings
