+++ Sources/Audio/Internals/Engine/EngineAudioUnit.swift
@@ -255,7 +255,7 @@ public class EngineAudioUnit: AUAudioUnit {
                 }
 
                 // Add render jobs for taps.
-                for tap in Tap.getTapsFor(node: node) {
+                for tap in TapRegistry.shared.getTapsFor(node: node) {
 
                     // We don't actually care about this output buffer. Perhaps
                     // there's a better way to express this?
